Finishing authentication flow

+++ pages/infrastructure/[name].js
@@ -5,12 +5,18 @@ import api from "../../utils/api";
 import { useRouter } from 'next/router';
 import { Section, SectionHeader, SectionBody } from "../../components/bootstrap/Section";
 import { BreadcrumbHeader, BreadcrumbItem } from "../../components/bootstrap/SectionBreadcrumb";
-import { Card, Row, Col, Button, ButtonGroup } from "react-bootstrap";
+import { Card, Row, Col, Button, ButtonGroup, Spinner } from "react-bootstrap";
 import swal from "sweetalert";
+import nookies from "nookies";
+import useSWR from "swr";
 
-const ShowStackPage = ({ stack }) => {
+const fetcher = url => api.get(url, {headers: { "Authorization": "Bearer " + nookies.get().token}}).then(res => res.data.data)
+
+const ShowStackPage = () => {
     const router = useRouter();
     const { name } = router.query;
+    const token = nookies.get().token;
+    const { data, error } = useSWR(name ? `/stacks/describe/${name}` : null, fetcher);
 
     const enableProtection = () => {
         swal({
@@ -22,7 +28,7 @@ const ShowStackPage = ({ stack }) => {
         })
             .then((willEnable) => {
                 if (willEnable) {
-                    api.get(`/stacks/update/${name}?protect=true`)
+                    api.get(`/stacks/update/${name}?protect=true`, {headers: { "Authorization": "Bearer " + token}})
                         .then((response) => {
                             swal({
                                 title: response.data.status,
@@ -48,7 +54,7 @@ const ShowStackPage = ({ stack }) => {
         })
             .then((willDisable) => {
                 if (willDisable) {
-                    api.get(`/stacks/update/${name}?protect=false`)
+                    api.get(`/stacks/update/${name}?protect=false`, {headers: { "Authorization": "Bearer " + token}})
                         .then((response) => {
                             swal({
                                 title: response.data.status,
@@ -74,7 +80,7 @@ const ShowStackPage = ({ stack }) => {
         })
             .then((willDelete) => {
                 if (willDelete) {
-                    api.delete("/stacks/" + name)
+                    api.delete("/stacks/" + name, {headers: { "Authorization": "Bearer " + token}})
                         .then((response) => {
                             swal({
                                 title: response.data.status,
@@ -115,21 +121,29 @@ const ShowStackPage = ({ stack }) => {
                                 <Card>
                                     <Card.Header><h4>Stack Information</h4></Card.Header>
                                     <Card.Body>
-                                        <b>Stack Id</b> <p>{stack[0].StackId}</p>
-                                        <b>Stack Name</b> <p>{stack[0].StackName}</p>
-                                        <b>Stack Creation</b> <p>{dateFormat(stack[0].CreationTime, "dd/mm/yyyy HH:MM:ss")}</p>
-                                        <b>Stack Status</b> <p>{stack[0].StackStatus}</p>
-                                        <b>Disable Rollback</b>
-                                        <p>{stack[0].DisableRollback ? "true" : "false"}</p>
-                                        <b>Enable Termination Protection</b>
-                                        <p>{stack[0].EnableTerminationProtection ? "true" : "false"}</p>
+                                        {!data ? (
+                                            <div className="text-center">
+                                                <Spinner animation="border" variant="primary" />
+                                            </div>
+                                        ) : (
+                                            <>
+                                                <b>Stack Id</b> <p>{data[0].StackId}</p>
+                                                <b>Stack Name</b> <p>{data[0].StackName}</p>
+                                                <b>Stack Creation</b> <p>{dateFormat(data[0].CreationTime, "dd/mm/yyyy HH:MM:ss")}</p>
+                                                <b>Stack Status</b> <p>{data[0].StackStatus}</p>
+                                                <b>Disable Rollback</b>
+                                                <p>{data[0].DisableRollback ? "true" : "false"}</p>
+                                                <b>Enable Termination Protection</b>
+                                                <p>{data[0].EnableTerminationProtection ? "true" : "false"}</p>
+                                            </>
+                                        )}
                                     </Card.Body>
                                     <Card.Footer>
                                         <ButtonGroup aria-label="Button Operation">
                                             <Button variant="warning" onClick={() => router.push({ pathname: '/infrastructure/edit/[name]', query: { name: name } })}>
                                                 <i className="fas fa-edit"></i> Update
                                             </Button>
-                                            {stack[0].EnableTerminationProtection ? (
+                                            {data && data[0].EnableTerminationProtection ? (
                                                 <Button variant="danger" onClick={() => disableProtection()}>
                                                     <i className="fas fa-unlock"></i> Disable Termination Protection
                                                 </Button>
@@ -153,23 +167,6 @@ const ShowStackPage = ({ stack }) => {
             </Layout>
         </>
     );
-}
-
-export async function getStaticPaths() {
-    const res = await api.get("/stacks")
-    const stacks = await res.data.data
-
-    const paths = stacks.map((stack) => ({
-        params: { name: stack.StackName },
-    }))
-
-    return { paths, fallback: false }
-}
-
-export async function getStaticProps({ params }) {
-    const res = await api.get(`/stacks/describe/${params.name}`)
-    const stack = await res.data.data
-    return { props: { stack } }
 }
 
 export default ShowStackPage;
